# Improve performance of BigInteger.ToString("x")