Benchmark:
```C#
using System;
using System.Numerics;
using BenchmarkDotNet.Attributes.Jobs;
using BenchmarkDotNet.Attributes;
using BenchmarkDotNet.Running;

[MemoryDiagnoser]
[InProcess]
public class Program
{
    public static void Main() => BenchmarkRunner.Run<Program>();

    private static BigInteger s_smallBigInt = new BigInteger(100);
    private static BigInteger s_largeBigInt = BigInteger.Parse(new string('9', 2000));

    [Benchmark]
    public void BigIntegerToString_Small()
    {
        BigInteger small = s_smallBigInt;
        for (int i = 0; i < 100_000; i++) small.ToString("x");
    }

    [Benchmark]
    public void BigIntegerToString_Large()
    {
        BigInteger large = s_largeBigInt;
        for (int i = 0; i < 1_000; i++) large.ToString("x");
    }
}
```

Before:
```
 |                   Method |     Mean |     Error |    StdDev |     Gen 0 | Allocated |
 |------------------------- |---------:|----------:|----------:|----------:|----------:|
 | BigIntegerToString_Small | 20.01 ms | 0.4187 ms | 0.6519 ms | 6843.7500 |  27.47 MB |
 | BigIntegerToString_Large | 29.62 ms | 0.3435 ms | 0.3045 ms | 8468.7500 |  33.91 MB |
```

After:
```
 |                   Method |     Mean |     Error |    StdDev |     Gen 0 | Allocated |
 |------------------------- |---------:|----------:|----------:|----------:|----------:|
 | BigIntegerToString_Small | 9.445 ms | 0.1153 ms | 0.0834 ms |  750.0000 |   3.05 MB |
 | BigIntegerToString_Large | 3.946 ms | 0.0603 ms | 0.0534 ms | 1000.0000 |   4.01 MB |
```

So for this small value, an ~2x throughput improvement, and for this large value, an ~8x throughput improvement, and for both, an ~9x reduction in allocation.

cc: @bartonjs